Split SphericalNucleus to a separate file
