+++ src/general/model_potential.h
@@ -5,6 +5,7 @@
 #include "ModelPotential.h"
 #include "PointNucleus.h"
 #include "RadialPotential.h"
+#include "SphericalNucleus.h"
 #include "GaussianNucleus.h"
 
 namespace helfem {
@@ -19,25 +20,6 @@ namespace helfem {
     } nuclear_model_t;
     /// Get nuclear model
     ModelPotential * get_nuclear_model(nuclear_model_t model, int Z, double Rrms);
-
-    /// Uniformly charged spherical nucleus
-    class SphericalNucleus : public ModelPotential {
-      /// Charge
-      int Z;
-      /// Size
-      double R0;
-    public:
-      /// Constructor
-      SphericalNucleus(int Z, double Rrms);
-      /// Destructor
-      ~SphericalNucleus();
-      /// Potential
-      double V(double r) const override;
-      /// Get R0
-      double get_R0() const;
-      /// Set R0
-      void set_R0(double R0);
-    };
 
     /// Thin hollow nucleus
     class HollowNucleus : public ModelPotential {
